Inspector Limits (#2406)
The inspector was slow when a nested container value had a large number
of children. For example, a record with an array field that had 10,000 items.

When the record was expanded, it would display the whole array on one line. 
All 10,000 items. Now this is limited to 15.

Now when you click that array to expand it, it was creating rows for all 
10,000 items. This was also slow. Now it only will show the first 100 items, 
with a button to load the next 100.

To do this, we keep track of the "page" that the user has loaded in the
 inspector state. We store it in a map with the value index path as the 
key, and the page number as the value. Each time the button is 
clicked, the page number gets incremented and the next page is 
rendered.

This applies to all container values.