Correlations (#2455)
This is exciting. We've heard from users who want to add custom correlation queries to brim. Before this PR, each correlation (zeek, suricata, md5) was implemented slightly differently and was hard-coded. This unifies the implementation and takes a huge step towards third-party correlations.

A significant change was extending the "Results" state slice. Now it can hold results for any number of queries, not just the main query. You provide an "id" whenever you set or select results.

I also created a simple api.query method on our BrimApi class. It cleans up after itself so that you don't have to think about it.

Finally, I've made a correlations/run() thunk that uses the above two features. Each correlation is given an id, the query is issued, and the results are saved to the Results state slice under that id.