Fix Detail Window (#2494)
The Log Detail window now works as intended. Double clicking a log in the results, or clicking the "open in new window" button in the detail right pane, will open the log detail window.

We used to serialize all of the state in the search window, then send it to the log detail window. Now we only send over the encoded zed value, and the url.

When the detail window boots up, it will set the url and value, and the window will behave properly.

Note: I had to make the sessionQueries state slice a global state slice. The log detail window needs access to the query that it came from. It uses the url to find it. The session queries need to be available outside of their originating window, so I added a "$" to the reducer and we are good to go.

I also created a concept called Operation and SpecialOperation.

An Operation is an action performed on the main process and is requested through ipc from the client. Using electron's ipc apis does not give any type safety to the arguments or channel names. This Operation wrapper gives us that and autocomplete.

And finally, a special operation is one that can return one-off values when a condition is met. We do this frequently in the app. For example, the main process will kick off a new window, then wait for that window to request it's specific data.

I'd love feedback on the api design and the names of the functions/methods/classes.

Also Mason, I've created another magic folder. Any exported operations that live in the electron/ops folder will automatically have ".listen(main)" called on them. See electron/run-op-listeners.ts. 😅

Fixes #2416