zealot: send empty Content-Type header in load request
Zealot's Client.load method sends a "Content-Type: application/json"
request header even though we want the Zed lake service to perform
format auto-detection for these requests.  Sending that header isn't a
problem right now because the load endpoint ignores the Content-Type
header and always does auto-detection, but brimdata/zed#4160 will change
the load endpoint to respect that header and do auto-detection only if
it's absent or empty.  In preparation for that, change Client.load to
send an empty Content-Type header so the service will continue to do
auto-detection..