Pool Folders (#2572)
Pools can now be grouped by a specified delimiter. The default is the "/" character.

The groups can be collapsed and expanded.

When a group is renamed, all of the pools under that group get renamed one by one. This can be improved when the backend can handle multiple requests without error. Then we can send many rename requests at once.

When a group is deleted, all of the pools under that group get deleted one by one.