Fix Those Buggy Windows (#2610)
The bug:

When creating a new window, the app took these steps:

1. Load the window
2. Setup a "closed" listener to delete the reference to the window

If you closed the window quickly before it finished loading, the first step would throw an error. The second step would never happen. This lead to us keep around a reference to a destroyed window.

Any other places in the code that tried to access anything on that window would get the "Object is Destroyed" message.

The fix is to do this:

1. Setup the "closed" listener
2. Load the window

Now, no matter what, when the window is closed, we will remove the reference to it. Just like the electron docs instruct: https://www.electronjs.org/docs/latest/api/browser-window#event-closed

If the load fails, it will throw an error. I've gone to all the places were windows are created and wrapped them in a try catch so that we get helpful error messages.