Use server-supplied audience in Auth0Client
Auth0Client uses a fixed audience value (https://lake.brimdata.io) in
the authorization request it sends to Auth0, limiting Auth0Client to use
with Auth0 APIs created with that value as their API identifier.  Change
Auth0Client to use the audience supplied by the lake service in the GET
/auth/methods response instead.  (The lake service's audience is
configured with the -auth.audience flag to "zed serve", added in
brimdata/zed#4356).