Zui with No Tabs Open (#2797)
In order to achieve this, we had to make these changes.

Remove the lake from the URL
The "url" for each tab was previously prefixed with /lake/:id. This had to be removed, and a migration created to migrate all the urls saved in the state. The routes have been updated to reflect this as well.

Store the current lake id in the state for each window.
A window will only be able to connect to one lake at a time. This is like how slack and vscode have designed their apps. Their "workspaces" are the equivalent of our "lakes". You can have multiple windows connected to different lakes though. Wrote migration to set the lake id to whatever the lake id in the last opened tab.

Remove the constraint of having 1 tab always open.

Create a view for when there are no tabs open. Right now it's just the big zed logo background image.

Assign the tabs to a lake. When switching lakes, the tabs for the previous lake will be saved and hidden. When switching back, they will be restored.

Extras:

I made the width and presence of the detail pane the same for every tab in a window. This keeps it from jumping around as you move through the tabs. Feels much better.
I styled the tabs slightly different.