Clone array before storing in redux (#2802)
When we put things in our Redux store, it "for our convenience" makes objects and arrays non-extensible. The problem here was that while that object was frozen, the zealot lib tried to add to it. If we make a copy before putting it in the store, our problems get solved.

This appeared with the s3 backed lake because of the latency. Also, this does not happen in production I don't think.