Session List (#3132)
Added a sessions pane in the left sidebar.

Whenever a new query tab is created, a new entry will be added to the list of sessions. When the tab is closed, the session entry will remain. When the session is clicked, it will either activate the associated tab if it exists, or it will restore that tab, along with it's history. You can right click a session to delete it. Deleting it will destroy the associated tab and history.

The display name of the tab is text of the most recent query in that sessions history.